Add TrackInfo/AudioInfo/VideoInfo serialisers https://bugs.webkit.org/show_bug.cgi?id=299169 rdar://160929565
Reviewed by Youenn Fablet.

We can also remove the unused RemoteAudioInfo/RemoteVideoInfo object
they are unusused and would have redundant following this commit.

No change in observable behaviours. The serialisers will be used in a
follow-up change.

* Source/WebCore/platform/TrackInfo.cpp:
(WebCore::TrackInfo::createWithVariant):
* Source/WebCore/platform/TrackInfo.h:
(WebCore::TrackInfo::TrackInfo):
(WebCore::VideoInfo::create):
(WebCore::VideoInfo::VideoInfo):
(WebCore::AudioInfo::AudioInfo):
(WebCore::AudioInfo::create):
* Source/WebKit/CMakeLists.txt:
* Source/WebKit/DerivedSources-input.xcfilelist:
* Source/WebKit/DerivedSources.make:
* Source/WebKit/GPUProcess/media/RemoteTrackInfo.h: Removed.
* Source/WebKit/GPUProcess/media/RemoteTrackInfo.serialization.in: Removed.
* Source/WebKit/Scripts/webkit/messages.py:
(headers_for_type):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/300399@main
