+++ Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in
@@ -9279,6 +9279,36 @@ header: <WebCore/TrackInfo.h>
     Text
 };
 
+header: <WebCore/TrackInfo.h>
+[CustomHeader, RefCounted, DisableMissingMemberCheck] struct WebCore::AudioInfo {
+    WebCore::FourCC codecName;
+    String codecString;
+    WebCore::TrackID trackID;
+    [NotSerialized] WebCore::TrackInfoTrackType m_type;
+    uint32_t rate;
+    uint32_t channels;
+    uint32_t framesPerPacket;
+    uint8_t bitDepth;
+    RefPtr<WebCore::SharedBuffer> cookieData;
+};
+
+header: <WebCore/TrackInfo.h>
+[CustomHeader, RefCounted, DisableMissingMemberCheck] struct WebCore::VideoInfo {
+    WebCore::FourCC codecName;
+    String codecString;
+    WebCore::TrackID trackID;
+    [NotSerialized] WebCore::TrackInfoTrackType m_type;
+    WebCore::FloatSize size;
+    WebCore::FloatSize displaySize;
+    uint8_t bitDepth;
+    WebCore::PlatformVideoColorSpace colorSpace;
+    RefPtr<WebCore::SharedBuffer> atomData;
+};
+
+[RefCounted, CreateUsing=fromVariant] struct WebCore::TrackInfo {
+    Variant<Ref<WebCore::AudioInfo>, Ref<WebCore::VideoInfo>> toVariant();
+};
+
 header: <WebCore/MediaSamplesBlock.h>
 [CustomHeader, Nested] class WebCore::MediaSamplesBlock::MediaSampleItem {
     MediaTime presentationTime;
